Replace json2sh with jq
Previously, we relied on updating all of our state from a json
file with after processing with the json2sh utilitiy. Instead,
this changes the script to extract only what it needs with jq.

This is needed as the Dell5k does not like the json2sh script.
Further this should make it more clear where these values
originate.