Run wait-pelion-identity job on the DELL 5k (#5)
* Reformat consistantly

 * Replaced all spacing with a consistant style of 2 spaces.
 * Reduced line lenth
 * Removed needless bangs (`!`)

* Replace absolute bashlib paths with variable

This should make it easier to change in the future
Replace absolute paths to executable with vairable
Use SCRIPT_DIR vairable where applicable

* Accept the wigwag root dir as the first argument

Previously, this scirpt was incompatible with snap packaging as it
would look for absolute paths that don't exist. Now, this script
accepts the path to the  `/wigwag` directory as the fist argument,
which should allow us to run this within a snap

* Consolidate bash `test` to use `[]` instead of `[[]]`

Turns out `==` is not allowed in `test`. Instead, `=` should
be used.

* Simplify grabbing edge-core status

* Use temp files and dirs for temporary stuff

Turns out you should probably not be putting these things onto disk

Previously, we would dump bits and pecies of stuff in the directory
containing these scripts. That sounds error prone to me. Instead,
we create a temporary directory or temporary files to hold the same
information.

* Replace json2sh with jq

Previously, we relied on updating all of our state from a json
file with after processing with the json2sh utilitiy. Instead,
this changes the script to extract only what it needs with jq.

This is needed as the Dell5k does not like the json2sh script.
Further this should make it more clear where these values
originate.

* Remove unused `cd` and `cp`

This script was doing a lot of `cd` within functions. This can
make it hard for humans to track the working directory, as these
functions modify global state.

* Specify the port as the second argument

* Accept the identity directory as the third argument

* Allow NODE_MODULES to be passed through the enviroment

Or, don't overwrite it unconditionally

* Add a bunch of ';'

* Pass the temp_certs dir to the js

* Pass script dir to js

* Pass idenity director to js

* Create dirs without errors

* Avoid changing directories

* Give create-dev-identity it's own cert dir

* Put "eeprom" next to identity

* Run create-new-eeprom from the identity dir

This should workaround it trying to open `./old_eeprom.json`
and failing to do so.

* Remove some calls to cd

* Reduce the number of identity generation steps to 1

* Move gateway address processing into JS

Turns out that `[[]]` did some things that `[]` is incapable of.
Particularly, it could match globs. Instead, I'm using JS to match
globs because it has a builtin regex matcher.

* Remove unused functions

* Generate service address in identity for edge-k8s

* adding config

adding a no dhcp config for the dell 5000 gateway build.

* Removed the use tput as there is no terminal when the script runs

* Removed the hardcoded NODE_PATH from the scripts

Co-authored-by: Jordan Rife <jrife0@gmail.com>
Co-authored-by: Mike Welsh <mike.welsh@arm.com>
Co-authored-by: Yash Goyal <Yash.Goyal@arm.com>